+++ packages/common/src/operations.graphql
@@ -116,167 +116,6 @@ mutation insertStocktake($input: InsertStocktakeInput!) {
   }
 }
 
-# query requisitions($params: RequisitionListParameters) {
-#   requisitions(params: $params) {
-#     __typename
-#     ... on RequisitionConnector {
-#       nodes {
-#         id
-#         comment
-#         orderDate
-#         theirReference
-#         requisitionNumber
-#         status
-#         otherPartyName
-#         otherPartyId
-#         colour
-#       }
-#       totalCount
-#     }
-
-#     ... on ConnectorError {
-#       __typename
-#       error {
-#         description
-#         ... on DatabaseError {
-#           __typename
-#           description
-#           fullError
-#         }
-#       }
-#     }
-#   }
-# }
-
-# mutation deleteCustomerRequisitions($ids: [DeleteRequestRequisitionInput!]) {
-#   batchRequestRequisition(deleteCustomerRequisitions: $ids) {
-#     __typename
-#     deleteRequestRequisitions {
-#       __typename
-#       id
-#     }
-#   }
-# }
-
-mutation updateCustomerRequisition(
-  $storeId: String!
-  $input: UpdateRequestRequisitionInput!
-) {
-  updateRequestRequisition(storeId: $storeId, input: $input) {
-    ... on RequisitionNode {
-      __typename
-      id
-    }
-  }
-}
-
-mutation insertCustomerRequisition(
-  $storeId: String!
-  $input: InsertRequestRequisitionInput!
-) {
-  insertRequestRequisition(input: $input, storeId: $storeId) {
-    ... on RequisitionNode {
-      __typename
-      id
-    }
-  }
-}
-
-query requisition($storeId: String!, $id: String!) {
-  requisition(id: $id, storeId: $storeId) {
-    __typename
-    ... on RequisitionNode {
-      __typename
-      id
-      # orderDate
-      # requisitionDate
-      comment
-      theirReference
-      type
-      requisitionNumber
-      # thresholdMOS
-      # maxMOS
-      status
-      otherPartyId
-      lines {
-        __typename
-        ... on RequisitionLineConnector {
-          totalCount
-          nodes {
-            id
-            # itemName
-            # itemCode
-            # itemUnit
-            itemId
-            # comment
-            # monthlyConsumption
-            # monthsOfSupply
-            supplyQuantity
-            # openingQuantity
-            # issuedQuantity
-            requestedQuantity
-            # receivedQuantity
-            # imprestQuantity
-            # previousQuantity
-            calculatedQuantity
-            # previousStockOnHand
-            # closingQuantity
-            # stockAdditions
-            # stockLosses
-            # expiredQuantity
-            # otherPartyClosingQuantity
-          }
-        }
-      }
-      otherParty {
-        __typename
-        ... on NameNode {
-          id
-          name
-          code
-          isCustomer
-          isSupplier
-        }
-      }
-    }
-  }
-}
-
-# mutation upsertCustomerRequisition(
-#   $deleteCustomerRequisitionLines: [DeleteCustomerRequisitionLineInput!]
-#   $insertCustomerRequisitionLines: [InsertCustomerRequisitionLineInput!]
-#   $updateCustomerRequisitionLines: [UpdateCustomerRequisitionLineInput!]
-#   $updateCustomerRequisitions: [UpdateCustomerRequisitionInput!]
-# ) {
-#   batchCustomerRequisition(
-#     deleteCustomerRequisitionLines: $deleteCustomerRequisitionLines
-#     insertCustomerRequisitionLines: $insertCustomerRequisitionLines
-#     updateCustomerRequisitionLines: $updateCustomerRequisitionLines
-#     updateCustomerRequisitions: $updateCustomerRequisitions
-#   ) {
-#     __typename
-#     updateCustomerRequisitions {
-#       __typename
-#       id
-#     }
-
-#     insertCustomerRequisitionLines {
-#       __typename
-#       id
-#     }
-
-#     deleteCustomerRequisitionLines {
-#       __typename
-#       id
-#     }
-
-#     updateCustomerRequisitionLines {
-#       __typename
-#       id
-#     }
-#   }
-# }
-
 query names(
   $key: NameSortFieldInput!
   $desc: Boolean
