+++ packages/requisitions/src/ResponseRequisition/DetailView/ContentArea.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 import { DataTable, useTranslation } from '@openmsupply-client/common';
-import { useCustomerRequisitionColumns } from './columns';
+import { useResponseRequisitionColumns } from './columns';
 import { useResponseRequisitionLines } from '../api';
 
 export const ContentArea: FC = () => {
   const { lines, onChangeSortBy, sortBy, pagination } =
     useResponseRequisitionLines();
-  const columns = useCustomerRequisitionColumns({ sortBy, onChangeSortBy });
+  const columns = useResponseRequisitionColumns({ sortBy, onChangeSortBy });
   const t = useTranslation('common');
 
   return (
